等级经验系统完成&&修复技能栏场景切换bug&&修复背包系统拾取bug

+++ MMO-SERVER/GameServer/Database/DbEntity.cs
@@ -30,7 +30,7 @@ namespace GameServer.Database
         public int Hp { get; set; } = 100;
         public int Mp { get; set; } = 100;
         public int Level { get; set; } = 1;
-        public int Exp { get; set; } = 0;
+        public long Exp { get; set; } = 0;
         public int SpaceId { get; set; }
         public int X { get; set; } = 0;
         public int Y { get; set; } = 0;
